requirements.txt
vkbottle==4.8.2
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
openai>=1.55
gspread==6.1.2
google-auth==2.29.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.9.5
pytz==2024.1
